Mouse cursor lock and toggle
During gameplay the mouse visibility is able to be toggled with the ESC
key, this will also lock the mouse at the same time. The script is in
there separately but doesn't seem to work so I just copied and pasted it
into the player script and it seems to work. Feel free to make it
neater.