Added Weapon Swap Functionality
You can switch between 2 weapons with TAB. Each weapon has its own
magazines but they use the same ammo, which would make sense with
paintball i guess.
Note that i changed the player prefab so that it doesnt include a gun
but instead and empty which is a weaponPlaceholder, this is where the
weapon is instantiated at, this has the advantage that if you pick extra
weapons up (future feature for sure) you dont have to edit the prefab
and can just add the new weapon as a child of the holder. Access the
weapon through player.currentWeapon.
Also i used some basic place holder weapons, consisting out of basic
cubes, and yes i know they look shit