Aim that actually works and Colourfull stuff
When the player shoots a raycast from the camera determines where the
player is aiming and then the bullet is rotated towards that point at
instantiation, so the bullets actually hit what youre aiming at...
A players colour is determined by his team colour, currently this is
being generated randomly but ideally would be picked from a list of
colours later on. Atm a players colour will look different from any
other players POV because they are not synced with the server. This
colour changes the colour of the players paintballs and the splatters
from those. In the future this should affect the players names as well.