Aim accuracy and name colour
When firing the playershoot script chooses a random point inside the
players crosshair and shoots there. this is further effected by the guns
accuracy. Testing gun 1 has an accuracy of 0 and will therefore shoot
randomly anywhere inside its crosshairs, meanwhile testing gun 2 has an
accuracy of 100 and will therefore always shoot right in the center of
it cross hair.

Also ive made it so the players name colour matches their splatter and
paintball colour, however as this is still chosen randomly it will look
a different colour for every player