Name Synched Repaired. Mouse only locks to screen after player loads.
The naming issue was called by my ignorance. I called the original name
call outside the loop i created for it to stay as a local player.
Moving the call inside that fixed the issue. All computers will have the
right name, and everyone else will have the right names as well.

Changed the player script to lock the mouse only after local player is
ready.