Scoreboard with kills and deaths for each player
This is synced accross all players, the only issue is that for some
reason whenever a client is killed its death function is called twice
which results in 2 deaths instead of 1, this doesnt occur for the host
though. Also this is implemented in the fort map but not the mine cos
they are different scenes.